Replaces color variables with colors map
Creates a color map in the _imports/_colors.scss and uses a color function
throughout the framework to generate background and color classes.  Allowing
custom colors to be used throughout the framework by adding a key to the
$colors map in _imports/_colors.scss

Dynamically adds color style classes to text, backgrounds, and buttons. Removes
individual color variables in favor of a color map.

Adds primary, second, and accent to color examples and adds new build

Bumps minor version 1.1.0